LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-reporting-check-report-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.

CONSOLE MESSAGE: ReferenceError: Can't find variable: trustedTypes

Harness Error (FAIL), message = ReferenceError: Can't find variable: trustedTypes
